+++ src/components/utils.ts
@@ -1,4 +1,3 @@
-import { Table } from '../models/misc'
 import { Notify } from 'quasar'
 
 export function success(msg: string, icon = 'check_circle') {
@@ -16,15 +15,12 @@ export function fail(msg: string) {
   })
 }
 
-export function getList(url: string, table: Table) {
-  table.loading = true
-  /*
-  this.$axios.get(url).then(res => {
-    if (res) {
-      table.data = res.data.data
-    }
-    table.loading = false
-  })*/
+export function status2color(status: number) {
+  if (status < 200) return 'accent'
+  else if (status < 300) return 'green'
+  else if (status < 400) return 'teal'
+  else if (status < 500) return 'orange'
+  else return 'red'
 }
 
 export function fmtTime(timestamp: number): string {