+++ src/models/misc.ts
@@ -1,7 +1,0 @@
-export interface Status {
-  code: number
-}
-
-export interface Table {
-  loading: boolean
-}